refactor: extract evaluate & get_concept logic into knowledge_graph/operations
